+++ frontend/src/components/SideBar/SideBarStreamerList/SideBarStreamer/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './SideBarStreamer';
