+++ frontend/src/components/SideBar/SideBarStreamerList/index.jsx
@@ -1,0 +1,1 @@
+export { defalut } from './SideBarStreamerList';
